feat: Add support for underlines
