+++ src/lib/markdown/rules.ts
@@ -1,11 +1,11 @@
 import markdownit from "markdown-it";
-import markPlugin from "markdown-it-mark";
-import placeholderPlugin from "./placeholders";
+import markPlugin from "./mark";
 import checkboxPlugin from "./checkboxes";
 import embedsPlugin from "./embeds";
 import breakPlugin from "./breaks";
 import tablesPlugin from "./tables";
 import noticesPlugin from "./notices";
+import underlinesPlugin from "./underlines";
 
 export default function rules({ embeds }) {
   return markdownit("default", {
@@ -15,8 +15,9 @@ export default function rules({ embeds }) {
     .use(embedsPlugin(embeds))
     .use(breakPlugin)
     .use(checkboxPlugin)
-    .use(markPlugin)
-    .use(placeholderPlugin)
+    .use(markPlugin({ delim: "==", mark: "mark" }))
+    .use(markPlugin({ delim: "!!", mark: "placeholder" }))
+    .use(underlinesPlugin)
     .use(tablesPlugin)
     .use(noticesPlugin);
 }
